requirements.txt
streamlit==1.25.0
faiss-cpu==1.7.4
sentence-transformers==2.2.2
numpy==1.24.3
pandas==2.0.2
plotly==5.15.0
huggingface-hub==0.18.1
